# docs/api/openapi/linksee-v1.yaml
openapi: 3.0.3
info:
  title: Linksee Teaching Collaboration API
  version: 1.0.0
  description: >
    Linksee 教学项目协作与过程评价平台的 P0 API schema。
    本文件与 docs/api/contracts/communication-contract-v1.md 保持一致，
    优先覆盖课程项目、阶段要求、小组讨论、阶段提交、教师反馈和课程看板。
servers:
  - url: /api/v1
    description: Linksee API v1
...
paths:
...
  /courses/{courseId}/members:
    get:
      tags:
        - Courses
      summary: 获取课程学生列表
      operationId: listCourseMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, withdrawn]
            default: active
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: 课程学生列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        status:
                          type: string
                          enum: [active, withdrawn]
                        joinedAt:
                          type: string
                          format: date-time
                        user:
                          type: object
                          properties:
                            id:
                              type: string
                            profile:
                              type: object
                              nullable: true
                              properties:
                                realName:
                                  type: string
                                accountNo:
                                  type: string
                                  nullable: true
                                avatarUrl:
                                  type: string
                                  nullable: true
                            studentProfile:
                              type: object
                              nullable: true
                              properties:
                                stuNo:
                                  type: string
                                grade:
                                  type: integer
                                adminClass:
                                  type: string
                  paging:
                    $ref: "#/components/schemas/Paging"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags:
        - Courses
      summary: 为课程添加单个学生
      operationId: addCourseMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
      responses:
        "201":
          description: 添加成功
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/members/batch:
    post:
      tags:
        - Courses
      summary: 为课程批量添加学生
      operationId: batchAddCourseMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userIds]
              properties:
                userIds:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: 批量添加成功
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /courses/{courseId}/members/{userId}:
    delete:
      tags:
        - Courses
      summary: 将学生移出课程
      operationId: removeCourseMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 移出成功
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
      description: 课程 ID。
...
  responses:
    Unauthenticated:
      description: 未认证或 token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: UNAUTHENTICATED
            message: Authentication token is missing or invalid.
            details: null
            requestId: trace_20260430_001
    Forbidden:
      description: 当前用户无权访问该资源或执行该操作
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: FORBIDDEN
            message: Current user is not allowed to perform this operation.
            details: null
            requestId: trace_20260430_002
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: NOT_FOUND
            message: Requested resource was not found.
            details: null
            requestId: trace_20260430_003
...
    ValidationFailed:
      description: 请求字段校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: VALIDATION_FAILED
            message: Request payload validation failed.
            details: null
            requestId: trace_20260430_005
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - code
        - message
      properties:
        ok:
          type: boolean
          enum: [false]
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - VALIDATION_FAILED
            - INTERNAL_ERROR
            - FORCE_CHANGE_PASSWORD
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
          nullable: true
...
    Paging:
      type: object
      required:
        - limit
        - offset
        - total
        - hasMore
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
        hasMore:
          type: boolean
...